refactor: move util methods to util class
Move bundle verification methods
to utility class.

closes #16617
